+++ app/src/main/java/com/example/myapplication_1/Adapters/AdapterWish.java
@@ -1,0 +1,119 @@
+package com.example.myapplication_1.Adapters;
+
+
+import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.drawable.Drawable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.myapplication_1.R;
+
+import ru.osety.amironlibrary.DrawableUtils;
+
+public class AdapterWish {
+
+    public static class AdapterGridViewMenu extends RecyclerView.Adapter<AdapterGridViewMenu.ViewHolder> {
+
+        private final ItemsMenu[] itemsMenu;
+        private final LayoutInflater layoutInflater;
+        private final Context context;
+
+        public AdapterGridViewMenu(AdapterGridViewMenu.ItemsMenu[] itemsMenu, Context context) {
+            this.itemsMenu = itemsMenu;
+            this.context = context;
+
+            layoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        }
+
+        @NonNull
+        @Override
+        public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
+
+            ViewGroup v = (ViewGroup) layoutInflater.inflate(R.layout.cell_wishes_model, viewGroup, false);
+            return new ViewHolder(v);
+
+        }
+
+        @Override
+        public void onBindViewHolder(@NonNull ViewHolder viewHolder, int i) {
+
+            final ItemsMenu _item = itemsMenu[i];
+
+            Drawable ic_background_xml = context.getResources().getDrawable(R.mipmap.icon_car);
+            float _dens = context.getResources().getDisplayMetrics().density;
+
+            Drawable ic_rout_color = DrawableUtils.setTintDrawable(ic_background_xml, _item.getColorBackground());
+
+            int _size = Math.round(_dens * 40);
+            Drawable _def_draw = context.getResources().getDrawable(_item.getImgResId());
+            Bitmap _bitmap = DrawableUtils.convertToBitmap(_def_draw, _size, _size);
+
+            //  viewHolder.img.setBackground( ic_rout_color );
+            //viewHolder.img.setScaleType(ImageView.ScaleType.CENTER);
+            //viewHolder.img.setImageBitmap( _bitmap );
+            viewHolder.desc.setText(_item.getStr());
+            viewHolder.cost.setText(_item.getCost());
+        }
+
+        @Override
+        public int getItemCount() {
+            return itemsMenu.length;
+        }
+
+        public static class ItemsMenu {
+
+
+            private @ColorInt
+            int colorBackgroundInt;
+            private int imgResId;
+            private String cost;
+            private String desc;
+
+            public ItemsMenu(int colorBackgroundRes, int imgResId, String cost, String desc) {
+                this.colorBackgroundInt = colorBackgroundRes;
+                this.imgResId = imgResId;
+                this.cost = cost;
+                this.desc = desc;
+            }
+
+            public int getImgResId() {
+                return imgResId;
+            }
+
+            public String getStr() {
+                return desc;
+            }
+
+            public String getCost() {
+                return cost;
+            }
+
+            public int getColorBackground() {
+                return colorBackgroundInt;
+            }
+        }
+
+
+        class ViewHolder extends RecyclerView.ViewHolder {
+
+            private final ImageView img;
+            private final TextView desc;
+            private final TextView cost;
+
+            public ViewHolder(@NonNull View itemView) {
+                super(itemView);
+                img = itemView.findViewById(R.id.img);
+                desc = itemView.findViewById(R.id.desc);
+                cost = itemView.findViewById(R.id.cost);
+            }
+        }
+    }
+}
