+++ app/src/main/res/layout/cell_travels_cost_change.xml
@@ -34,7 +34,7 @@
             android:textSize="15dp"/>
 
         <SeekBar
-            android:id="@+id/seekBar"
+            android:id="@+id/ll_travel_cost_change_seekBar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="#FCFCFC"
@@ -44,13 +44,13 @@
             android:thumb="@drawable/custom_thumb" />
 
         <TextView
-            android:id="@+id/txtView"
+            android:id="@+id/ll_travel_cost_change_txtView"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textAlignment="center"
             android:paddingLeft="5dp"
             android:layout_centerHorizontal="true"
-            android:text="0 ₽"
+            android:text="100 ₽"
             android:gravity="center_horizontal"
             android:textSize="17sp" />
